@@ -16,41 +16,47 @@
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/stack": "^6.3.20",
     "@shopify/flash-list": "1.4.3",
-    "@shopify/react-native-skia": "0.1.196",
     "@tanstack/react-query": "^5.14.2",
-    "@types/jest": "^29.5.11",
     "axios": "^1.6.5",
-    "expo": "^49.0.21",
+    "expo": "^50.0.1",
+    "expo-application": "~5.8.3",
     "expo-auth-session": "~5.0.2",
     "expo-build-properties": "~0.8.3",
     "expo-checkbox": "~2.4.0",
+    "expo-constants": "~15.4.2",
     "expo-crypto": "~12.4.1",
     "expo-image-picker": "~14.3.2",
+    "expo-linking": "~6.2.1",
+    "expo-router": "~3.4.1",
     "expo-secure-store": "~12.3.1",
-    "expo-splash-screen": "~0.20.5",
-    "expo-status-bar": "~1.6.0",
+    "expo-splash-screen": "~0.26.1",
+    "expo-status-bar": "~1.11.1",
     "expo-updates": "~0.18.18",
     "expo-web-browser": "~12.3.2",
-    "jest": "^29.2.1",
-    "jest-expo": "~49.0.0",
-    "nativewind": "^2.0.11",
+    "nativewind": "^4.0.1",
     "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.49.3",
-    "react-native": "0.72.6",
+    "react-native": "0.73.2",
     "react-native-gesture-handler": "~2.12.0",
-    "react-native-reanimated": "~3.3.0",
-    "react-native-safe-area-context": "4.6.3",
+    "react-native-reanimated": "~3.6.0",
+    "react-native-safe-area-context": "4.8.2",
+    "react-native-screens": "~3.29.0",
     "react-native-svg": "13.9.0",
     "react-native-url-polyfill": "^2.0.0",
+    "react-native-web": "~0.19.6",
     "zod": "^3.22.4",
-    "zustand": "^4.4.7"
+    "zustand": "^4.4.7",
+    "tailwindcss": "3.4.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.11",
+    "jest": "^29.2.1",
+    "jest-expo": "~49.0.0",
     "@babel/core": "^7.20.0",
     "@openapitools/openapi-generator-cli": "^2.7.0",
     "@testing-library/react-native": "^12.4.3",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "@types/react": "~18.2.14",
     "@types/react-test-renderer": "^18.0.7",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
@@ -61,8 +67,8 @@
     "eslint-plugin-react-native": "^4.1.0",
     "prettier": "^3.1.1",
     "react-test-renderer": "^18.2.0",
-    "tailwindcss": "3.3.2",
-    "typescript": "^5.1.3"
+    "@types/react": "~18.2.45",
+    "typescript": "^5.3.0"
   },
   "private": true
 }
