@@ -1,9 +1,11 @@
 module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'],
-    plugins: [
+    presets: [
+      ['babel-preset-expo', { jsxImportSource: 'nativewind' }],
       'nativewind/babel',
+    ],
+    plugins: [
       [
         'module-resolver',
         {
