@@ -1,7 +1,10 @@
-import { styled } from 'nativewind';
+import { type PropsWithChildren } from 'react';
 import { View } from 'react-native';
 
-export const ExpensesUpsert = styled(
-  View,
-  'flex-1 items-center justify-center bg-app-primary-100',
-);
+export const ExpensesUpsertContainer = ({ children }: PropsWithChildren) => {
+  return (
+    <View className="flex-1 items-center justify-center bg-app-primary-100">
+      {children}
+    </View>
+  );
+};
